lookup 1000 groups when attempting to cache them so we are not restricted to 100 w/ burnettk
